Only try to delete revision directory if it exists

+++ lib/spontaneous/publishing/steps/create_revision_directory.rb
@@ -14,7 +14,7 @@ module Spontaneous::Publishing::Steps
 
     # This is the reason for the existance of this step: cleaning up when it's gone wrong
     def rollback
-      FileUtils.rm_r(path)
+      FileUtils.rm_r(path) if File.exist?(path)
     end
 
     def path
